debugging

+++ plugin-non-client-call-control/src/components/CustomEndCallButton.js
@@ -7,23 +7,29 @@ import {
   templates,
   ContentFragment,
 } from '@twilio/flex-ui';
-
 import ConferenceService from '../services/ConferenceService';
 import FlexState from '../states/FlexState';
 
 class CustomEndCallButton extends React.PureComponent {
+  constructor(props) {
+    super(props);
+    this.state = { disabled: false };
+  }
   handleClick = async () => {
     console.debug('*** Custom End Call button clicked');
-    const { task } = this.props;
-    console.debug('PROPS OBJ: ', this.props);
+    const { task, manager } = this.props;
 
     //await ConferenceService.cancelTask(task.taskSid);
     const workerParticipant = FlexState.getLocalParticipantForTask(task);
     const { callSid: participantCallSid } = workerParticipant;
-    await ConferenceService.removeParticipant(
-      task?.conference?.conferenceSid,
-      participantCallSid
-    );
+    if (this.state.disabled === false) {
+      this.setState({ disabled: true });
+      const removeResults = await ConferenceService.removeParticipant(
+        task?.conference?.conferenceSid,
+        participantCallSid
+      );
+      console.debug('REMOVE RESULTS: ', removeResults);
+    }
   };
   render() {
     const { task, theme } = this.props;
